dispatcher/requirements.txt
functions-framework>=3.5
google-cloud-compute>=1.14
